// src/app/core/Model/categories.model.ts
import { SelectModel } from "src/app/component/search-bar/select.model";

export interface Categories {
	bakery: SelectModel[];
	fruitAndVegetables: SelectModel[];
	meatAndFish: SelectModel[];
	drinks: SelectModel[];
	kitchen: SelectModel[];
	specialNutrition: SelectModel[];
	baby: SelectModel[];
	pharmacy: SelectModel[];
}